Paiment multiple : la procédure d'archivage ne doit pas supprimer les entrées de email_fields_map_reference qui concernent les paiements multiples encore en cours.
Pour mémoire, pour regénérer les email_fields_map_reference des paiements multiples supprimés par la procédure d'archivage :

BEGIN;

INSERT INTO email_fields_map_reference (
    id,
    field2,
    field1,
    mail,
    reference,
    version,
    pay_evt_montant,
    date_created,
    billing_address1,
    billing_city,
    billing_country_code,
    billing_firstname,
    billing_lastname,
    billing_zip_code,
    shoppingcart_total_quantity,
    is_payed
)
SELECT
    nextval('hibernate_sequence'),
    ptl.field2,
    ptl.field1,
    ptl.mail,
    ptl.reference,
    ptl.version,
    ptl.pay_evt_montant,
    ptl.transaction_date,
    ptl.billing_address1,
    ptl.billing_city,
    ptl.billing_country_code,
    ptl.billing_firstname,
    ptl.billing_lastname,
    ptl.billing_zip_code,
    ptl.shoppingcart_total_quantity,
    (ptl.erreur = '00000')
FROM pay_transaction_log ptl
WHERE ptl.pay_evt_montant IN (
    SELECT id FROM pay_evt_montant WHERE paiement_multiple_montant1 IS NOT NULL
)
AND ptl.reference NOT IN (
    SELECT reference FROM email_fields_map_reference
);

COMMIT;

+++ src/main/java/org/esupportail/pay/services/ArchiveService.java
@@ -17,11 +17,16 @@
  */
 package org.esupportail.pay.services;
 
+import java.time.LocalDate;
+import java.util.Date;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import jakarta.annotation.Resource;
 import jakarta.persistence.TypedQuery;
 
+import org.esupportail.pay.domain.PayEvtMontant;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.esupportail.pay.dao.EmailFieldsMapReferenceDaoService;
@@ -66,13 +71,29 @@ public class ArchiveService {
 		if(enabled) {
 			log.debug("removeOldTmpEmailFieldsMapReference called");
 			List<EmailFieldsMapReference> emailFieldsMapReferences2remove = emailFieldsMapReferenceDaoService.findOldEmailFieldsMapReferences(oldDays4emailFieldsMapReference);
+			Set<EmailFieldsMapReference> emailFieldsMapReferences2notRemoveBecausePaimentMultiple = new HashSet<>();
 			for(EmailFieldsMapReference emailFieldsMapReference : emailFieldsMapReferences2remove) {
-				TypedQuery<ScienceConfReference> q = scienceConfReferenceDaoService.findScienceConfReferencesByEmailFieldsMapReference(emailFieldsMapReference);
-				if(!q.getResultList().isEmpty()) {
-					scienceConfReferenceDaoService.remove(q.getSingleResult());
+				PayEvtMontant payEvtMontant = emailFieldsMapReference.getPayEvtMontant();
+				if(payEvtMontant.getPaiementMultiple_kind().isEmpty()) {
+					TypedQuery<ScienceConfReference> q = scienceConfReferenceDaoService.findScienceConfReferencesByEmailFieldsMapReference(emailFieldsMapReference);
+					if (!q.getResultList().isEmpty()) {
+						scienceConfReferenceDaoService.remove(q.getSingleResult());
+					}
+					emailFieldsMapReferenceDaoService.remove(emailFieldsMapReference);
+				} else {
+					// multiple paiement : emailFieldsMapReference should not be removed before all paiement is OK
+					LocalDate lastDate = payEvtMontant.getPaiementMultiple_last_date();
+					if(lastDate.plusDays(oldDays4emailFieldsMapReference).isAfter(LocalDate.now())) {
+						emailFieldsMapReferenceDaoService.remove(emailFieldsMapReference);
+					}  else {
+						emailFieldsMapReferences2notRemoveBecausePaimentMultiple.add(emailFieldsMapReference);
+					}
 				}
-				emailFieldsMapReferenceDaoService.remove(emailFieldsMapReference);
 			}
+			if(emailFieldsMapReferences2notRemoveBecausePaimentMultiple.size()>0) {
+				log.debug(emailFieldsMapReferences2notRemoveBecausePaimentMultiple.size() + " old emailFieldsMapReferences not removed because used in multiple paiements");
+			}
+			emailFieldsMapReferences2remove.removeAll(emailFieldsMapReferences2notRemoveBecausePaimentMultiple);
 			if(emailFieldsMapReferences2remove.size()>0) {
 				log.info(emailFieldsMapReferences2remove.size() + " old emailFieldsMapReferences removed");
 			}
